refactor(lib): add credentials chain for s3Client (#6)
* refactor(lib): add logging of s3 vars

Signed-off-by: Kenton Duprey <kentonduprey@icloud.com>

* feat(lib): update getS3Client to use async/await

This commit updates the `getS3Client` function to use async/await syntax for retrieving the S3
client. It replaces the previous implementation that used a synchronous approach. This change
improves the efficiency and readability of the code.

Signed-off-by: Kenton Duprey <kentonduprey@icloud.com>

* fix(upload): update getS3Client with more logs for better readability

Signed-off-by: Kenton Duprey <kentonduprey@icloud.com>

* feat(upload): update getS3Client with improved credentials handling

This commit updates the `getS3Client` function in the `upload/server-actions.ts` file. It
improves the handling of S3 client credentials by introducing a check for the `VERCEL`
environment variable. If both `VERCEL` and `NEXT_PRIVATE_UPLOAD_AWS_ROLE_ARN` are present,
it uses AWS OIDC authentication credentials. Otherwise, it falls back to using the
`NEXT_PRIVATE_UPLOAD_ACCESS_KEY_ID` and `NEXT_PRIVATE_UPLOAD_SECRET_ACCESS_KEY` environment
variables. This change enhances the flexibility and security of the S3 client authentication
process.

Signed-off-by: Kenton Duprey <kentonduprey@icloud.com>

* chore(web): update dev script port to 3106

* refactor(upload): improve getS3Client credentials handling and logging

Signed-off-by: Kenton Duprey <kentonduprey@icloud.com>

* chore(lib): update AWS SDK dependencies to latest versions

Signed-off-by: Kenton Duprey <kentonduprey@icloud.com>

* refactor(upload): improve getS3Client logging

* refactor(upload): update getS3Client with improved credentials handling and logging

* refactor(upload): remove redundant logger in getS3Client

---------

Signed-off-by: Kenton Duprey <kentonduprey@icloud.com>